+++ CHANGELOG.md
@@ -13,7 +13,7 @@ This project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.htm
 - `powpow.grep` will no longer accept an empty `pattern` string, raising
   `ValueError` instead.
 - `powpow.GrepResult` now holds the string against which the match was
-  performed in its `.string` attribute.
+  performed in its `.input` attribute.
 - Return values of `powpow.GrepResult`'s `__str__` and `__repr__` methods, and
   `matches` property, are now cached.