+++ CHANGELOG.md
@@ -7,6 +7,10 @@ This project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.htm
 
 ## [Unreleased]
 ### Added
+Nothing yet
+
+## [0.0.4]
+### Added
 - Basic implementation of `powpow.cat` and `powpow.CatResult`.
 - `powpow.GrepResult` is now actually importable from the top-level namespace.
 - `powpow.GrepResult` is now guaranteed to have all of the methods the `str`
@@ -58,6 +62,7 @@ First version of the library.
 - The most basic, dashed of version of PowPow Grep utility.
 
 [Unreleased]: https://github.com/mrmino/powpow/compare/v0.0.3...HEAD
+[0.0.4]: https://github.com/mrmino/powpow/compare/v0.0.3...v0.0.4
 [0.0.3]: https://github.com/mrmino/powpow/compare/v0.0.2...v0.0.3
 [0.0.2]: https://github.com/mrmino/powpow/compare/v0.0.1...v0.0.2
 [0.0.1]: https://github.com/mrmino/powpow/releases/tag/v0.0.1